📝 #74 Doku im Code für Leveldatenhaltung geschrieben 🤡 #74 Unter großen Schmerzen ein Beispiellevel in die Datenbank eingepflegt 💥 Datenbank wird beim Start wieder über die Data.sql befüllt und bei Programm-Exit gedroppt 🍱 Beispiel-3D-Modelle eingefügt

+++ src/main/java/de/hsrm/mi/swt/rheinmainadventure/repositories/RaumMobiliarRepository.java
@@ -1,6 +1,5 @@
 package de.hsrm.mi.swt.rheinmainadventure.repositories;
 
-import de.hsrm.mi.swt.rheinmainadventure.entities.Raum;
 import de.hsrm.mi.swt.rheinmainadventure.entities.RaumMobiliar;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -10,5 +9,13 @@ import java.util.List;
 @Repository
 public interface RaumMobiliarRepository extends JpaRepository<RaumMobiliar, Long> {
 
-    List<RaumMobiliar> findAllByRaum_RaumId(int raumId);
+
+    /**
+     * Findet sämtliches Mobiliar, das sich in einem Raum befindet.
+     *
+     * @param raumId is die Raum-ID, zu der das Mobiliar nachgeschlagen werden soll
+     * @return Eine Liste von RaumMobiliar-Objekten, aus der man über die Mobiliar-ID dann das Mobiliar enthält
+     * sowie die X/Y-Position im Level
+     */
+    List<RaumMobiliar> findAllByRaum_RaumId(Long raumId);
 }
